requirements.txt
# Langachain dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

openai==1.6.1
# Faiss CPU was used tdur t
faiss-cpu==1.7.4

sentence-transformers==2.2.2
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

python-dotenv==1.0.0
tqdm==4.66.1
pytest==7.4.3
pytest-asyncio==0.21.1
gunicorn==21.2.0
python-multipart==0.0.6